Detangle inet keys and jp Henkan/ Muhenkan keys
When support for inet keys was added to the keyboard driver
back in the early 2000 none of the developers thought of
Japanese 106 key keybards which have two extra keys generating
the scan codes 0x79 and 0x7d.
So for keys prefixed with the scan code 0x60 which were not
remapped to anything else the line:
 *scanCode += 0x78;
was added.
Thus keys generating the scan codes 0xe0 0x01 and 0xe0 0x03
will overlap with the Henkan and Muhenkan keys on Japanese
keyboards.
A workaround was introduced to xkeyboard-config which makes
sure the keycodes 0x79 and 0x7d provide the Henkan and Muhenkan
keysyms. As a consequence users might loose some multimedia keys
on Japanese keyboards.
A different solution would be to either remap keycodes 0x79
and 0x7d or keycodes 0xe0 0x01 and 0xe0 0x03 into the range of
unused (X-)keycodes above 0xfb.
There are downsides with either solution: the legacy keyboard
driver and the xkb files would have to be fixed in sync. Otherwise
Japanese users would loose their Henkan and Muhenkan keys (in
the first case) or some multimedia keys would get lost.
The present patch does the latter which seems to be cleaner codewise.

Signed-off-by: Egbert Eich <eich@suse.de>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>